LPS-44956 move to module
