+++ modules/portal/portal-sso-ntlm/src/com/liferay/portal/sso/ntlm/NtlmManager.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.ntlm;
+package com.liferay.portal.sso.ntlm;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.StringPool;
