+++ modules/portal/portal-sso-ntlm/src/com/liferay/portal/sso/ntlm/msrpc/NetlogonIdentityInfo.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.ntlm.msrpc;
+package com.liferay.portal.sso.ntlm.msrpc;
 
 import jcifs.dcerpc.UnicodeString;
 import jcifs.dcerpc.ndr.NdrBuffer;
